Rename kotlin io's based capature results to the original class name
