+++ include-build/roborazzi-core/src/commonMain/kotlin/com/github/takahirom/roborazzi/CaptureResults.kt
@@ -1,5 +1,6 @@
 package com.github.takahirom.roborazzi
 
+import kotlinx.io.files.Path
 import kotlinx.serialization.ContextualSerializer
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
@@ -24,12 +25,11 @@ import kotlinx.serialization.json.intOrNull
 import kotlinx.serialization.json.long
 import kotlinx.serialization.json.longOrNull
 import kotlinx.serialization.modules.SerializersModule
-import java.io.File
 
 @Serializable
-data class CaptureResults(
+data class CaptureResults2(
   @SerialName("summary")
-  val resultSummary: ResultSummary,
+  val resultSummary: ResultSummary2,
   @SerialName("results")
   val captureResults: List<CaptureResult>
 ) {
@@ -69,7 +69,7 @@ data class CaptureResults(
           id = key,
           contents = buildString {
             captureResults.groupBy { it.contextData[key]?.toString() ?: "undefined" }
-              .toSortedMap(Comparator<String> { value1, value2 ->
+              .toSortedMap { value1, value2 ->
                 // show "undefined" and "null" at the end
                 if (value1 == "undefined" || value2 == "null") {
                   1
@@ -78,7 +78,7 @@ data class CaptureResults(
                 } else {
                   value1.compareTo(value2)
                 }
-              })
+              }
               .forEach { (value, results) ->
                 append(buildTable(value, value, results, reportDirectoryPath))
               }
@@ -111,10 +111,10 @@ data class CaptureResults(
     images: List<CaptureResult>,
     reportDirectoryPath: String
   ): String {
-    fun File.pathFrom(reportDirectoryPath: String): String {
-      val reportDirectory = File(reportDirectoryPath)
-      val relativePath = relativeTo(reportDirectory)
-      return relativePath.path
+    fun Path.pathFrom(reportDirectoryPath: String): String {
+      val reportDirectory = Path(reportDirectoryPath)
+      val relativePath =  relativeTo(reportDirectory)
+      return relativePath.toString()
     }
     if (images.isEmpty()) return ""
     return buildString {
@@ -166,18 +166,18 @@ data class CaptureResults(
       ignoreUnknownKeys = true
       classDiscriminator = "#class"
       serializersModule = SerializersModule {
-        contextual(File::class,
-          object : KSerializer<File> {
+        contextual(Path::class,
+          object : KSerializer<Path> {
             override val descriptor: SerialDescriptor =
               PrimitiveSerialDescriptor("FileSerializer", PrimitiveKind.STRING)
 
-            override fun serialize(encoder: Encoder, value: File) {
+            override fun serialize(encoder: Encoder, value: Path) {
               encoder.encodeString(value.absolutePath)
             }
 
-            override fun deserialize(decoder: Decoder): File {
+            override fun deserialize(decoder: Decoder): Path {
               val path = decoder.decodeString()
-              return File(path)
+              return Path(path)
             }
           }
         )
@@ -185,18 +185,19 @@ data class CaptureResults(
       }
     }
 
-    fun fromJsonFile(inputPath: String): CaptureResults {
-      val jsonElement = json.parseToJsonElement(File(inputPath).readText())
+    fun fromJsonFile(inputPath: String): CaptureResults2 {
+      val string = KotlinIo.readText(Path(inputPath))
+      val jsonElement = json.parseToJsonElement(string)
       return json.decodeFromJsonElement(jsonElement)
     }
 
-    fun fromJson(jsonString: JsonObject): CaptureResults {
+    fun fromJson(jsonString: JsonObject): CaptureResults2 {
       return json.decodeFromJsonElement(jsonString)
     }
 
-    fun from(results: List<CaptureResult>): CaptureResults {
-      return CaptureResults(
-        resultSummary = ResultSummary(
+    fun from(results: List<CaptureResult>): CaptureResults2 {
+      return CaptureResults2(
+        resultSummary = ResultSummary2(
           total = results.size,
           recorded = results.count { it is CaptureResult.Recorded },
           added = results.count { it is CaptureResult.Added },
@@ -238,4 +239,8 @@ object AnySerializer : KSerializer<Any> {
       else -> throw IllegalArgumentException("Unknown type: ${input::class.qualifiedName}")
     }
   }
+}
+
+fun <K : Comparable<K>, V> Map<out K, V>.toSortedMap(comparator: Comparator<in K>): Map<K, V> {
+  return this.toList().sortedWith(compareBy(comparator) { it.first }).toMap()
 }
